caffe-master/python/requirements.txt
scipy==1.0.0
six==1.11.0
numpy==1.14.1
scikit_image==0.13.1
Pillow==5.1.0
protobuf==3.5.2.post1
pydot==1.2.4
pydotplus==2.0.2
skimage==0.0
